requirements.txt
flask
flask-migrate 
flask-sqlalchemy
sqlalchemy-serializer
flask-restful
flask-cors
flask-bcrypt
